# src/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contact Book API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Це API для керування контактами з авторизацією, поштою та аватарками.\n"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://https://nodejs-hw-mongodb-lilya.onrender.com"
    }
  ],
...
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "Отримати всі контакти користувача",
        "description": "Повертає список усіх контактів, які належать користувачу",
        "operationId": "getAllContacts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Номер сторінки (пагінація)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Кількість елементів на сторінці"
          },
          {
            "name": "favorite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Фільтр за улюбленими"
          }
        ],
        "responses": {
          "200": {
            "description": "Список контактів",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований запит"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "post": {
          "tags": [
            "contacts"
          ],
          "summary": "Створити новий контакт",
          "description": "Додає новий контакт для авторизованого користувача",
          "operationId": "createContact",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../components/schemas/NewContact.yaml"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Контакт успішно створено",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "../../components/schemas/Contact.yaml"
                  }
                }
              }
            },
            "400": {
              "description": "Помилка валідації"
            },
            "401": {
              "description": "Неавторизований запит"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      }
    },
    "/contacts/{contactId}": {
      "get": {
        "200": {
          "description": "Успішна відповідь",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../../components/schemas/Contact.yaml"
              }
            }
          }
        },
        "tags": [
          "contacts"
        ],
        "summary": "Отримати контакт за ID",
        "operationId": "getContactById",
        "description": "Отримання одного контакту за його унікальним ідентифікатором",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": null
      },
      "patch": {
        "patch": {
          "tags": [
            "contacts"
          ],
          "summary": "Оновити контакт за ID",
          "description": "Оновлює контакт за ID авторизованого користувача",
          "operationId": "updateContact",
          "parameters": [
            {
              "name": "contactId",
              "in": "path",
              "required": true,
              "description": "Ідентифікатор контакту",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../components/schemas/NewContact.yaml"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Контакт оновлено",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "../../../components/schemas/Contact.yaml"
                  }
                }
              }
            },
            "400": {
              "description": "Некоректні дані"
            },
            "401": {
              "description": "Неавторизований запит"
            },
            "404": {
              "description": "Контакт не знайдено"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      },
      "delete": {
        "delete": {
          "tags": [
            "contacts"
          ],
          "summary": "Видалити контакт за ID",
          "description": "Видаляє контакт із бази даних",
          "operationId": "deleteContact",
          "parameters": [
            {
              "name": "contactId",
              "in": "path",
              "required": true,
              "description": "Ідентифікатор контакту",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Контакт успішно видалено"
            },
            "401": {
              "description": "Неавторизований запит"
            },
            "404": {
              "description": "Контакт не знайдено"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Contact": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "email",
          "phone",
          "favorite",
          "owner"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "favorite": {
            "type": "boolean"
          },
          "owner": {
            "type": "string"
          }
        }
      },
...
    }
  }
}